tss: Add new calls
- Local policy (non persistent)
- recoveryOS
- recoveryOS Local policy (persistent)

The local policy TSS request differs slightly between the one for the
restore process and the recoveryOS. The one for the restore process is
meant to be used only once, and the one for the recovery os is meant to
be loaded from disk, and thus has a different request. See the
Ap,LocalBoot parameter.